# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CSE 341 Board Game Library API",
    "version": "1.0.0",
    "description": "A REST API for a board game collection and the play sessions logged against it, backed by MongoDB Atlas. Built for CSE 341 Web Services, project 2.\n\nTwo collections: **games** (nine fields) and **sessions** (six). A session references its game by `gameId`, so creating one requires a game that already exists, and deleting a game also deletes its sessions.\n\nEvery POST and PUT validates its whole body and answers `400` with an `errors` array listing every problem found, not just the first one."
  },
  "servers": [
    {
      "url": "/",
      "description": "This server (localhost or Render)"
    }
  ],
...
  "paths": {
...
    "/games/{id}": {
      "get": {
        "tags": [
          "Games"
        ],
        "summary": "Get one game by its MongoDB ObjectId",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The game's MongoDB ObjectId (24 hex characters)",
            "example": "000000000000000000000000"
          }
        ],
        "responses": {
          "200": {
            "description": "The matching game",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameWithId"
                }
              }
            }
          },
          "400": {
            "description": "The id is not a valid ObjectId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No game has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "The database could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Games"
        ],
        "summary": "Replace an existing game",
        "description": "Replaces the whole game, so send all nine fields. The same validation rules as POST apply.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The game's MongoDB ObjectId (24 hex characters)",
            "example": "000000000000000000000000"
          }
        ],
        "responses": {
          "204": {
            "description": "Game updated. There is no response body."
          },
          "400": {
            "description": "The id is not a valid ObjectId, or a field is missing or badly formatted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No game has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "The database could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Game"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Games"
        ],
        "summary": "Delete a game and all of its play sessions",
        "description": "Sessions reference a game by id, so deleting a game also deletes its sessions. Leaving them behind would orphan them, pointing at a game that no longer exists.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The game's MongoDB ObjectId (24 hex characters)",
            "example": "000000000000000000000000"
          }
        ],
        "responses": {
          "204": {
            "description": "Game deleted. There is no response body."
          },
          "400": {
            "description": "The id is not a valid ObjectId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No game has that id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "The database could not be reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Game": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "required": {
            "type": "array",
            "example": [
              "title",
              "publisher",
              "yearPublished",
              "minPlayers",
              "maxPlayers",
              "playTimeMinutes",
              "complexity",
              "categories",
              "ownerEmail"
            ],
            "items": {
              "type": "string"
            }
          },
          "properties": {
            "type": "object",
            "properties": {
              "title": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Wingspan"
                  }
                }
              },
              "publisher": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "example": {
                    "type": "string",
                    "example": "Stonemaier Games"
                  }
                }
              },
              "yearPublished": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 2019
                  },
                  "minimum": {
                    "type": "number",
                    "example": 1900
                  }
                }
              },
              "minPlayers": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 1
                  },
                  "minimum": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "maxPlayers": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 5
                  },
                  "minimum": {
                    "type": "number",
                    "example": 1
                  },
                  "description": {
                    "type": "string",
                    "example": "Must be greater than or equal to minPlayers."
                  }
                }
              },
              "playTimeMinutes": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  },
                  "example": {
                    "type": "number",
                    "example": 70
                  },
                  "minimum": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "complexity": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "enum": {
                    "type": "array",
                    "example": [
                      "light",
                      "medium",
                      "heavy"
                    ],
                    "items": {
                      "type": "string"
                    }
                  },
                  "example": {
                    "type": "string",
                    "example": "medium"
                  }
                }
              },
              "categories": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "minItems": {
                    "type": "number",
                    "example": 1
                  },
                  "example": {
                    "type": "array",
                    "example": [
                      "strategy",
                      "engine-building"
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "ownerEmail": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "email"
                  },
                  "example": {
                    "type": "string",
                    "example": "reese@example.com"
                  }
                }
              }
            }
          }
        }
      },
      "GameWithId": {
        "type": "object",
        "properties": {
          "allOf": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Game"
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "A short summary of what went wrong."
                  },
                  "example": {
                    "type": "string",
                    "example": "Invalid game."
                  }
                }
              },
              "errors": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "description": {
                    "type": "string",
                    "example": "Present on validation failures only: one entry per problem found in the body."
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "example": {
                    "type": "array",
                    "example": [
                      "\"title\" is required and must be a non-empty string.",
                      "\"maxPlayers\" must be greater than or equal to \"minPlayers\"."
                    ],
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}